Create OCPBaselineRedfishService.json
Initial draft for a Redfish Service baseline profile.  This profile aims to create a baseline set of requirements for any Redfish-managed device specified by OCP.  Therefore, this profile does not include any requirements that rely on underlying hardware features or components, but only specifies properties implemented solely in the management controller's firmware.  Other OCP projects can then include this profile as a requirement in topic or hardware-specific profiles, while keeping both a consistent baseline of service functionality, and removing the duplication of effort.